Add node and list classes and also some methods
The node class constructor returns an object with three fields: value,
reference to the previous one to the next node. The list class
co-constructor forms the head and tail fields of the list.
The lengthFromTail method moves from the tail of the list until the
reference to the previous element is empty, counting and returning the
number of nodes. The lengthFromHead method works similarly, only from
the head, I did it in order to compare the results of these methods and
determine the correctness of other methods.
The append method takes values, the tape type is checked, then the tail
node becomes penultimate, and the tail field is replaced by a new node
that has a link to the penultimate, if the list head is empty, the
current node takes its role and the function ends the next element
becomes a link to the tail.
The insert method takes the value and position number in the list. We
start to run from the head to the location of the node, if the number
is greater than the number of nodes we see an error message. If the
searched place is the end of the list, then we call the append method,
otherwise we create a new node, assign it a link to neighboring nodes,
and neighboring nodes - to it. If this is the beginning of the list,
then change the header field with this element.