Add a delete method
The method takes the sequence number of the node. The algorithm for
finding the node by sequence number is similar to this operation in
the insert method. Then save the links to the next and previous
objects, then change their links to the next and previous,
respectively. If one of them is not, then the other is given a
reference to null and make it head or tail, respectively. Finally,
return the value of the deleted element.