Add clone and reverse methods
The first method creates a new list and, starting at the head of the
current list, adds nodes with values to the end of the new list using
the append method as long as nodes exist. Finally, the method returns
a new list.
The second method creates a new list and, like the previous method,
sorts the values of the nodes in the current list and adds the
corresponding nodes to the new one, but starting with the tail. At
the end, the corresponding fields of the new list are written in the
head and tail fields of the current list.