Add findFirst and findLast methods
In the first method, we go through all the nodes in the list until
the value of the node is equal to the desired one. If no match is
found, -1 is returned.
In the second method we create a temporary variable and also sort all
nodes. If the value matches, then write the node index in the time
variable, then continue the search. Finally, return the last value of
the time variable.