Add clear and extend methods
The first method clears the values of the head and tail fields of the
current list. Therefore, all links to this list are lost.
The second method goes through all the nodes in the transmitted list
and adds new nodes with the appropriate values to the current list
using the append method.